gather subteam and member row types
